Python 3.12 Focus and Dependency Updates
We now support only Python 3.12.
We also changed the build process to use virtual environments.
This means that a Makefile would be hard to use, because we need to set up and tear down the virtual environment script-wide.
So we use a Bash build script make.sh instead.
This needs to be supported by the build process.

We also version-stepped all dependencies to the latest versions.
This required some further updates in the code, because some static type checkers find stuff that they previously could not find.